requirements.txt
boto3
botocore
celery==5.3.1
click==8.1.3
Flask==2.3.1
flask_sqlalchemy==3.0.3
flask_cors==4.0.0
langchain~=0.3.14
openai
python-dotenv~=1.0.1
Werkzeug==2.3.1
invoke==2.2.0
pinecone-client==3.2.2
pypdf==3.15.4
watchdog==3.0.0
tiktoken
uuid==1.30
langfuse==1.0.18
backoff==2.2.1
redis==5.0.0
pydantic
pyboxen
langchain-community~=0.3.14
langchain-core~=0.3.29
langchain-openai~=0.2.14
chromadb
langchain-chroma